+++ lib/db/src/schema/index.ts
@@ -1,3 +1,4 @@
 export * from "./entries";
 export * from "./giveaways";
 export * from "./referralRewards";
+export * from "./betaSignups";